Move c++-based pattern miner in xpattern-miner folder
