+++ opencog/learning/xpattern-miner/HandleTree.cc
