+++ opencog/learning/xpattern-miner/HandleTree.h
