+++ opencog/learning/xpattern-miner/Valuations.cc
