+++ opencog/learning/xpattern-miner/Valuations.h
