+++ opencog/learning/xpattern-miner/XPatternMiner.cc
