+++ opencog/learning/xpattern-miner/XPatternMiner.h
