+++ tests/learning/xpattern-miner/CMakeLists.txt
@@ -1,2 +1,9 @@
-ADD_SUBDIRECTORY (c++-based)
-ADD_SUBDIRECTORY (ure-based)
+ADD_CXXTEST(XPatternMinerUTest)
+TARGET_LINK_LIBRARIES(XPatternMinerUTest
+  xpatternminer
+)
+
+ADD_CXXTEST(UREPatternMinerUTest)
+TARGET_LINK_LIBRARIES(UREPatternMinerUTest
+  xpatternminer
+)
