--- tests/learning/xpattern-miner/c++-based/CMakeLists.txt
@@ -1,4 +1,0 @@
-ADD_CXXTEST(XPatternMinerUTest)
-TARGET_LINK_LIBRARIES(XPatternMinerUTest
-  xpatternminer
-)
