--- tests/learning/xpattern-miner/ure-based/CMakeLists.txt
@@ -1,4 +1,0 @@
-ADD_CXXTEST(UREPatternMinerUTest)
-TARGET_LINK_LIBRARIES(UREPatternMinerUTest
-  xpatternminer
-)
